refactor(frontend): tooltip component config
Config tooltip components for each layer ID in `src/config/interaction-groups`, so that we can edit new layers in one place.
